wasm: fix building the wasm backend with the new `int` behaving like `isize`.

+++ examples/wasm_codegen/add.v
@@ -4,8 +4,8 @@ fn main() {
 	mut m := wasm.Module{}
 	mut func := m.new_function('add', [.i32_t, .i32_t], [.i32_t])
 	{
-		func.local_get(0)
-		func.local_get(1)
+		func.local_get(i32(0))
+		func.local_get(i32(1))
 		func.add(.i32_t)
 	}
 	m.commit(func, true) // `export: true`
