+++ examples/wasm_codegen/factorial.v
@@ -4,7 +4,7 @@ fn main() {
 	mut m := wasm.Module{}
 	mut fac := m.new_function('fac', [.i64_t], [.i64_t])
 	{
-		fac.local_get(0)
+		fac.local_get(i32(0))
 		fac.eqz(.i64_t)
 		bif := fac.c_if([], [.i64_t])
 		{
@@ -13,10 +13,10 @@ fn main() {
 		fac.c_else(bif)
 		{
 			{
-				fac.local_get(0)
+				fac.local_get(i32(0))
 			}
 			{
-				fac.local_get(0)
+				fac.local_get(i32(0))
 				fac.i64_const(1)
 				fac.sub(.i64_t)
 				fac.call('fac')
